+++ src/components/users/LoginForm.tsx
@@ -1,0 +1,75 @@
+import { Link, useNavigate } from "react-router-dom";
+import { useAuth } from "../../contexts/AuthContext";
+import { useState } from "react";
+
+export default function LoginForm() {
+      const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
+  const { login } = useAuth();
+  const navigate = useNavigate();
+
+  const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    const success = await login(email, password);
+
+    if (success) {
+      console.log('로그인 성공');
+      navigate('/'); // Redirect on successful login
+    } else {
+      console.log('로그인 실패: 사용자 정보가 일치하지 않습니다.');
+    }
+  };
+
+  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: { name, value },
+    } = e;
+
+    if (name === 'email') {
+      setEmail(value);
+    }
+
+    if (name === 'password') {
+      setPassword(value);
+    }
+
+  };
+  return     <div className="page flex flex-col items-center justify-center mt-6 px-[15px]">
+      <h1 className="text-4xl mb-3">Sign in</h1>
+      <p className="text-[#5cb85c] mb-4 hover:text-[#1E822A] hover:underline">
+        <Link to="/register">Need an account?</Link>
+      </p>
+      <form className="w-full" onSubmit={onSubmit}>
+        <div className="">
+          <input
+            className="px-6 py-3 text-xl w-full text-[#55595c] border rounded mb-4"
+            type="text"
+            name="email"
+            id="email"
+            onChange={onChange}
+            placeholder="Email"
+            value={email}
+          />
+        </div>
+        <div>
+          <input
+            className="px-6 py-3 text-xl w-full text-[#55595c] border rounded mb-4"
+            type="password"
+            name="password"
+            id="password"
+            onChange={onChange}
+            placeholder="Password"
+            value={password}
+          />
+        </div>
+        <div className="flex justify-end">
+          <button
+            type="submit"
+            className="px-6 py-3 text-xl rounded text-white bg-[#5cb85c] select-none hover:bg-[#1E822A]"
+          >
+            Sign in
+          </button>
+        </div>
+      </form>
+    </div>
+}
